update code BE

+++ be-spring-boot/src/main/java/com/springboot/appspringboot/repository/TagRepository.java
@@ -1,11 +1,8 @@
 package com.springboot.appspringboot.repository;
 
-import com.springboot.appspringboot.entity.Tag;
+import com.springboot.appspringboot.model.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
 @Repository
-public interface TagRepository extends JpaRepository<Tag, Integer> {
+public interface TagRepository extends JpaRepository<Tag, Long> {
 }
